Extract RPC access logging
